Money donation + worker fixes + documentation

+++ EchoProject.Application/Requests/Projects/GoalRequest.cs
@@ -3,9 +3,9 @@ namespace EchoProject.Application.Requests.Projects
     public record GoalRequest
     (
         string Title,
-        long TargetAmount,
-        long CurrentAmount,
-        long? CostPerUnit,
+        decimal TargetAmount,
+        decimal CurrentAmount,
+        decimal? CostPerUnit,
         List<Guid>? VendorIds,
         Guid GoalTypeId
     );
